# APICL.Client/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "APICL.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Image/images": {
      "get": {
        "tags": [
          "Image"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ImageObjInfo"
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/Image/images/{guid}/info": {
      "get": {
        "tags": [
          "Image"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "guid",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ImageObjInfo"
            }
          },
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/Image/images/{guid}/remove": {
      "delete": {
        "tags": [
          "Image"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "guid",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "boolean"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/Image/images/empty/{width}/{height}": {
      "post": {
        "tags": [
          "Image"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "width",
            "required": true,
            "type": "integer",
            "format": "int32",
            "default": 1920
          },
          {
            "in": "path",
            "name": "height",
            "required": true,
            "type": "integer",
            "format": "int32",
            "default": 1080
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/ImageObjInfo"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/Image/images/upload": {
      "post": {
        "tags": [
          "Image"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "copyGuid",
            "type": "boolean",
            "default": true
          },
          {
            "in": "formData",
            "name": "file",
            "type": "file"
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/ImageObjInfo"
            }
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/Image/images/{guid}/download/{format}": {
      "get": {
        "tags": [
          "Image"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "guid",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "format",
            "required": true,
            "type": "string",
            "default": "bmp"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/Image/images/{guid}/image64/{format}": {
      "get": {
        "tags": [
          "Image"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "guid",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "format",
            "required": true,
            "type": "string",
            "default": "bmp"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ImageData"
            }
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ImageData": {
      "type": "object",
      "properties": {
        "guid": {
          "format": "uuid",
          "type": "string"
        },
        "base64": {
          "type": "string"
        },
        "width": {
          "format": "int32",
          "type": "integer"
        },
        "height": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "ImageObjInfo": {
      "type": "object",
      "properties": {
        "guid": {
          "format": "uuid",
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "filePath": {
          "type": "string"
        },
        "height": {
          "format": "int32",
          "type": "integer"
        },
        "width": {
          "format": "int32",
          "type": "integer"
        },
        "bitdepth": {
          "format": "int32",
          "type": "integer"
        },
        "channels": {
          "format": "int32",
          "type": "integer"
        },
        "sizeInBytes": {
          "format": "int64",
          "type": "integer"
        },
        "pointer": {
          "format": "int64",
          "type": "integer"
        },
        "onHost": {
          "type": "boolean"
        },
        "entry": {
          "type": "string"
        },
        "lastProcessingTime": {
          "type": "string"
        },
        "lastLoadingTime": {
          "type": "string"
        },
        "errorInfo": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    }
  }
}